% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_pop.R
\name{agg_pop_q}
\alias{agg_pop_q}
\title{Core of agg_pop not exported. A function that aggregates the immunization population tallies of DRC per organization level using built-in 2025 population dataset}
\usage{
agg_pop_q(var, group, month = 1, pops)
}
\arguments{
\item{var}{Column name of the variable containing the population numbers. For the pops dataset monthly targets immunization use naissance_mois_ajust or survivant_mois_ajust. For surveillance use population_totale}

\item{group}{The organization level orgunitlevel1 for country, orgunitlevel2 for province, orgunitlevel3 for district and orgunitlevel6 for antennes}

\item{month}{A numeric value between 1 and 12 to specify the number of months for immunization targets}

\item{pops}{A dataframe with population data}
}
\value{
A dataframe of aggregated populations
}
\description{
Core of agg_pop not exported. A function that aggregates the immunization population tallies of DRC per organization level using built-in 2025 population dataset
}
