% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antennes}
\alias{antennes}
\title{A dataset containing DRC's health antennas and health districts}
\format{
A data frame with 519 observations and 3 variables:
\describe{
\item{province}{character; provinces}
\item{antenne}{character; health antennas}
\item{zonesante}{character; health zones}
}
}
\source{
Programme Élargie de Vaccination de RDC.
}
\usage{
antennes
}
\description{
A dataset containing DRC's health antennas and health districts
}
\keyword{datasets}
