% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_pop.R
\name{clean_dhis_names}
\alias{clean_dhis_names}
\title{A function that removes DRC DHIS prefixes and suffixes}
\usage{
clean_dhis_names(
  data,
  column,
  has_prefix = 2,
  suffix = "(Province|Zone de Sant[eé]|Aire de Sant[eé]|Centre de Sant[eé])"
)
}
\arguments{
\item{data}{The dataset containing variables to be cleaned}

\item{column}{The column containing values to be cleaned espcially organization levels of DRC EPI}

\item{has_prefix}{An integer specifying the number of characters of the prefix to be removed. Set to 0 if no prefix}

\item{suffix}{A character value that specifies the suffix to be removed}
}
\value{
A dataframe with cleaned names
}
\description{
A function that removes DRC DHIS prefixes and suffixes
}
\examples{

datafile <- system.file(c("pop2025.RDS"), package = "pevrdc")

vacc <- readRDS(datafile)
clean_dhis_names(vacc, column = orgunitlevel2, has_prefix = 2, suffix = "Province")

clean_names_dhis(vacc$orgunitlevel2,  has_prefix = 2,
                 suffix = "(Province|Zone de Sant[e\u00e9]|Aire de Sant[e\u00e9]|Centre de Sant[e\u00e9])")
}
