% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_pop.R
\name{clean_dhis_names_q}
\alias{clean_dhis_names_q}
\title{Core of clean_dhis_names not exported.#'}
\usage{
clean_dhis_names_q(
  data,
  column,
  has_prefix = 2,
  suffix = "(Province|Zone de Sant[eé]|Aire de Sant[eé]|Centre de Sant[eé])"
)
}
\arguments{
\item{data}{The dataset containing variables to be cleaned}

\item{column}{The column containing values to be cleaned espcially organization levels of DRC EPI}

\item{has_prefix}{An integer specifying the number of characters of the prefix to be removed. Set to 0 if no prefix}

\item{suffix}{A character value that specifies the suffix to be removed}
}
\value{
A dataframe with cleaned names
}
\description{
Core of clean_dhis_names not exported.#'
}
