% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_pop.R
\name{clean_names_dhis}
\alias{clean_names_dhis}
\title{A function that removes DRC's DHIS 2 prefixes and suffixes  from a vector}
\usage{
clean_names_dhis(
  x,
  has_prefix = 2,
  suffix = "(Province|Zone de Sant[eé]|Aire de Sant[eé]|Centre de Sant[eé])"
)
}
\arguments{
\item{x}{A character vector to be cleaned especially organization levels of DRC EPI. More convenient to use within other functions}

\item{has_prefix}{An integer specifying the number of characters of the prefix to be removed. Set to 0 if no prefix}

\item{suffix}{A character value that specifies the suffix to be removed}
}
\value{
A character vector
}
\description{
A function that removes DRC's DHIS 2 prefixes and suffixes  from a vector
}
