% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_accent.R
\name{correct_accent}
\alias{correct_accent}
\title{A function to remove accented characters from all the character values of a data frame}
\usage{
correct_accent(data, pattern = "[é]", replacement = "e")
}
\arguments{
\item{data}{Dataframe, usually the DHIS2 immunization dataframe}

\item{pattern}{A character value to match or a regular expression. Default "é|é"}

\item{replacement}{A character value to replace the pattern. Default "e"}
}
\value{
A dataframe with values corrected
}
\description{
A function to remove accented characters from all the character values of a data frame
}
\examples{
datafile <- system.file(c("vacc_data.RDS"), package = "pevrdc")

vacc <- readRDS(datafile)
correct_accent(vacc, pattern =  "[\u00C3\u00A9]", replacement = "e")
}
