% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outbreak.R
\name{district_epi}
\alias{district_epi}
\title{This is function is wrapper around is_outbreak to identify districts with an outbreak.}
\usage{
district_epi(
  df,
  district,
  wk,
  status,
  nwk = 4,
  sorted = FALSE,
  pre_yr = FALSE,
  year = NULL
)
}
\arguments{
\item{df}{A dataset with a weekly number of cases or a logical indication of a case detected}

\item{district}{A variable containing districts names}

\item{wk}{A numeric variable specifying the successive week numbers}

\item{status}{A logical or numeric variable with either indication of case recorded or the number of cases}

\item{nwk}{A numeric value that specifies the window frame for the assessment of consecutive logical values of x}

\item{sorted}{A logical value to specify whether the dataset is ordered per week and district}

\item{pre_yr}{A logical value to specify whether dataset contains data from the previous year}

\item{year}{A numeric variable containing the year of the data}
}
\value{
A dataframe with an additional logical variable epi_status
}
\description{
This is function is wrapper around is_outbreak to identify districts with an outbreak.
}
\examples{

df <- data.frame(district = c(rep("A", 4), rep("B", 5)),
                 week = c(1:4, 1:5), 
                 status = c(TRUE, FALSE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE))

district_epi(df, district, week, status, 3, sorted = TRUE)

}
