% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outbreak.R
\name{is_outbreak}
\alias{is_outbreak}
\title{This is function returns TRUE after n successive TRUE and FALSE otherwise. This is a convenient function to classify areas with outbreak following consecutive
positive cases of diseases recorded within n weeks.}
\usage{
is_outbreak(x, nwk = 4)
}
\arguments{
\item{x}{an ordered logical vector}

\item{nwk}{a numeric value that specifies the window frame for the assessment of consecutive logical values of x}
}
\value{
A logical vecor of similar length with x
}
\description{
This is function returns TRUE after n successive TRUE and FALSE otherwise. This is a convenient function to classify areas with outbreak following consecutive
positive cases of diseases recorded within n weeks.
}
\examples{

is_outbreak(c(TRUE, FALSE, TRUE, TRUE, TRUE, TRUE, TRUE), 4)

}
