% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop2025}
\alias{pop2025}
\title{A dataset containing DRC's 2025 target immunization populations at health districts levels}
\format{
A data frame with 519 observations and 6 variables:
\describe{
\item{orgunitlevel1}{character; Country}
\item{orgunitlevel2}{character; Province}
\item{orgunitlevel3}{character; Zone de santé (health district)  }
\item{naissance_mois_admin}{numeric; monthly number of livebirths}
\item{survivants_mois_admin}{numeric; monthly number of surviving children}
\item{naissance_mois_ajust}{numeric; monthly number of livebirths ajusted}
\item{survivant_mois_ajust}{numeric; monthly number of surviving children ajusted}
\item{orgunitlevel6}{character; antenne}
}
}
\source{
Programme Élargie de Vaccination de RDC.
}
\usage{
pop2025
}
\description{
A dataset containing DRC's 2025 target immunization populations at health districts levels
}
\keyword{datasets}
