% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape}
\alias{shape}
\title{A dataset with the DRC shapefile at health zone level, necessitate the sf package to load.}
\format{
A data frame with 519 observations and 5 variables:
\describe{
\item{pays}{character}
\item{province}{character}
\item{antenne}{character}
\item{zonesante}{character}
\item{SHAPE}{multipolygon}
}
}
\source{
GRID-3.
}
\usage{
shape
}
\description{
A dataset with the DRC shapefile at health zone level, necessitate the sf package to load.
}
\keyword{datasets}
