% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacc_data}
\alias{vacc_data}
\title{A dataset with simulated DHIS 2 imuundata containing DRC's health antennas and health districts
Please refer to DHIS 2 for a complete codebook of the data}
\format{
A data frame with 1000 observations and 225 variables:
}
\source{
Programme Élargie de Vaccination de RDC.
}
\usage{
vacc_data
}
\description{
A dataset with simulated DHIS 2 imuundata containing DRC's health antennas and health districts
Please refer to DHIS 2 for a complete codebook of the data
}
\keyword{datasets}
