% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_sums.R
\name{vaccine_sums}
\alias{vaccine_sums}
\title{A function to compute the cummulative monthly tally of vaccinated per organization levels using DHIS2 exported data}
\usage{
vaccine_sums(data, vaccine, level, month = 1, accent = TRUE)
}
\arguments{
\item{data}{Dataframe, usually the DHIS2 immunization dataframe}

\item{vaccine}{Column name of the vaccine of interest. Should be one of the following: BCG, DTC1, DTC3, VPO0, VAA, VAR1, VAR2, VPI1, VPI2, Pneumo13, Rota1, Rota2, Rota3, VAP1, VAP2, VAP3, VAP4, DTC2, Pneumo13_2, VPI 2, VPO1, VPO2, VPO3 & TD2}

\item{level}{The column name or vector of column names of the aggregation level of interest. Should be any of Country, Province, District, or Antenne}

\item{month}{An integer between 1 to 12 that specifies the cummulative period of months to aggregate date}

\item{accent}{A logical to specify whether the dataset values contains accented characters. Set to TRUE as default}
}
\value{
A dataframe with aggregated sums of vaccinated between
}
\description{
A function to compute the cummulative monthly tally of vaccinated per organization levels using DHIS2 exported data
}
\examples{
datafile <- system.file(c("vacc_data.RDS"), package = "pevrdc")

vacc <- readRDS(datafile)
vaccine_sums(vacc, DTC1, District, 2)
}
