% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_antennes.R
\name{add_antennes}
\alias{add_antennes}
\title{A function that adds DRC antennes to a dataset not having them}
\usage{
add_antennes(data, province, zone, rename_antenne = TRUE)
}
\arguments{
\item{data}{A dataset with no variable containing antennes, usually DHIS 2 datasets}

\item{province}{A Column with province}

\item{zone}{A column with health zones}

\item{rename_antenne}{A logical to indicate renaming antennes following DHIS 2 standard}
}
\value{
A dataframe with a variable with antennes
}
\description{
A function that adds DRC antennes to a dataset not having them
}
