% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_antennes.R
\name{add_antennes}
\alias{add_antennes}
\title{A function that adds DRC antennes to a dataset not having them}
\usage{
add_antennes(data, dhis = TRUE, zone_sante = NULL, rename_antenne = TRUE)
}
\arguments{
\item{data}{A dataset with no variable containing antennes, usually DHIS 2 datasets}

\item{dhis}{A boolean to indicate whether the dataset is from the DHIS 2, set to TRUE as default}

\item{zone_sante}{A variable with health zones names if data are not from DHIS 2, set to NULL as default}

\item{rename_antenne}{A logical to indicate renaming antennes following DHIS 2 standard}
}
\value{
A dataframe with a variable with antennes
}
\description{
A function that adds DRC antennes to a dataset not having them
}
